backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
redis
kafka-python
celery
sqlalchemy
minio